requirements.txt
uvicorn==0.23.2
fastapi~=0.100.1
python-dotenv==1.0.0
PyYAML==6.0.1
oracledb==2.5.1
pydantic~=2.0.3
coverage~=7.2.7
pytest
pytest-asyncio
requests==2.31.0
oci~=2.111.0
tzdata
